Consolidated everything into one file.
I am toying with the idea of just have 1 file to download.
This way, we can ensure that everything works "smoothly."
Although it creates one giant soup of a setup file...
perhaps we may have less difficulty debugging this way?
I wanted to test this a few times.
I am probably totally wrong.
So wrong.
Just wanted it on the table.